backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
yt-dlp==2025.3.31
opencv-python-headless==4.11.0.86
pytesseract==0.3.13
anthropic==0.52.0
httpx==0.28.1
pillow==11.1.0
numpy==2.2.4
pydantic==2.10.6
beautifulsoup4==4.13.3
